+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: bkiziler <bkiziler@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 11:26:30 by bkiziler          #+#    #+#             */
-/*   Updated: 2022/11/16 14:12:41 by bkiziler         ###   ########.fr       */
+/*   Updated: 2022/11/17 13:39:10 by bkiziler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void			*ft_memmove(void *dst, const void *src, size_t len);
 void			*ft_memcpy(void *dst, const void *src, size_t n);
 void			ft_bzero(void *s, size_t n);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
-unsigned int	ft_strlcpy(char *dst, const char *src, size_t dstsize);
+size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
 int				ft_tolower(int c);
 int				ft_toupper(int c);
 char			*ft_strchr(const char *s, int c);
